meirong/requirements.txt
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Werkzeug==1.0.1
SQLAlchemy==1.3.19
PyJWT==1.7.1
python-dotenv==0.14.0